System Prefs: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@39271 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/prefs/screenmode/prefs.c
@@ -87,7 +87,7 @@ BOOL Prefs_ImportFH(BPTR fh)
 
                 if (error < 0)
                 {
-                    printf("Error: ReadChunkBytes() returned %d!\n", error);
+                    printf("Error: ReadChunkBytes() returned %d!\n", (int)error);
                 }
                 else
                 {
@@ -101,13 +101,13 @@ BOOL Prefs_ImportFH(BPTR fh)
             }
             else
             {
-                printf("ParseIFF() failed, returncode %d!\n", error);
+                printf("ParseIFF() failed, returncode %d!\n", (int)error);
                 success = FALSE;
             }
         }
         else
         {
-            printf("StopChunk() failed, returncode %d!\n", error);
+            printf("StopChunk() failed, returncode %d!\n", (int)error);
             success = FALSE;
         }
 
@@ -163,7 +163,7 @@ BOOL Prefs_ExportFH(BPTR fh)
 
             if (error != 0) // TODO: We need some error checking here!
             {
-                printf("error: PushChunk() = %d ", error);
+                printf("error: PushChunk() = %d ", (int)error);
             }
 
             error = WriteChunkBytes(handle, &saveprefs, sizeof(struct ScreenModePrefs));
@@ -171,7 +171,7 @@ BOOL Prefs_ExportFH(BPTR fh)
 
             if (error != 0) // TODO: We need some error checking here!
             {
-                printf("error: PopChunk() = %d ", error);
+                printf("error: PopChunk() = %d ", (int)error);
             }
 
             // Terminate the FORM
